+++ common/test/src/Model/tst_BrushRegression.cpp
@@ -24,6 +24,7 @@
 #include "IO/DiskIO.h"
 #include "IO/NodeReader.h"
 #include "IO/TestParserStatus.h"
+#include "Matchers.h"
 #include "Model/Brush.h"
 #include "Model/BrushBuilder.h"
 #include "Model/BrushFace.h"
